autohdr_client_exe/requirements.txt
customtkinter==5.2.2
requests==2.31.0
sseclient-py==1.8.0
pytest==8.0.0
tkinterdnd2-universal==1.7.3
python-dotenv==1.0.1
